The EndPoint PUT and DELETE are finalized
